All routes added, verification for api key and jwt token added, administrator verification for account methods added

+++ controllers/BaseController.js
@@ -6,30 +6,51 @@ class BaseController {
     getOne = async (req, res) => {
         let id = req.params.id;
         let found = await this.service.findOne({ where: {id:id}});
-        return res.status(200).send(found);
+        return res.status(200).json(found);
     }
 
     getAll = async (req, res) => {
         let all = await this.service.findAll();
-        return res.status(200).send(all);
+        return res.status(200).json(all);
     }
 
     create = async (req, res) => {
-        let created = await this.service.create(req.body);
-        return res.status(201).send(created);
+        let created;
+        try {
+            created = await this.service.create(req.body);
+            return res.status(201).json({ created, message: 'Utworzono'});
+        } catch (error) {
+            return res.status(400).json(error);
+        }
     }
 
     update = async (req, res) => {
         let id = req.params.id;
-        // TODO: Check if id exists
-        let updated = await this.service.update(req.body, { where : { id: id }});
-        return res.status(200).send(updated);
+        try {
+            let found = await this.service.findOne({ where: {id:id}});
+            if(found) {
+                let updated = await this.service.update(req.body, { where : { id: id }});
+                return res.status(200).json({ numberOfUpdatedRows: updated, message: 'Zaktualizowano' });
+            }
+            return res.status(404).json({ message: 'Nie znaleziono' });
+        } catch (error) {
+            return res.status(400).json(error);
+        }
     }
 
     delete = async (req, res) => {
         let id = req.params.id;
-        let destroyed = await this.service.destroy({ where: { id: id }});
-        return res.status(200).send(destroyed);
+        try {
+            let found = await this.service.findOne({ where: {id:id}});
+            if(found) {
+                let destroyed = await this.service.destroy({ where: { id: id }});
+                return res.status(200).json({ numberOfDestroyedRows: destroyed, message: 'Usunięto' });
+            }
+            return res.status(404).json({ numberOfDestroyedRows: 0, message: 'Nie znaleziono' });
+        }
+        catch (error) {
+            return res.status(400).json(error);
+        }
     }
 }
